Break pagination out into separate component
